The shapes now use float to draw the data
The shapes are implmented in such a way that they can accept float nature whilist passing the data

+++ Classassignment/Circle.cs
@@ -7,40 +7,40 @@ using System.Drawing;
 
 namespace Classassignment
 {
-    public  class Circle :Shapes
+    public class Circle : Shapes
     {
         int radius;
         Pen p;
-        SolidBrush  b;
+        SolidBrush b;
         /// <summary>
         /// This method  returns the int value and pen 
         /// </summary>
         /// <param name="data"></param>
         /// <param name="pen"></param>
-        public void getValue (int[] data , Pen pen)
+        public void getValue(float[] data, Pen pen)
         {
-             radius = data[0];
-                p = pen;
+            radius = (int)data[0];
+            p = pen;
         }
         /// <summary>
         /// This method returns the int data and solidbrush
         /// </summary>
         /// <param name="data"></param>
         /// <param name="brush"></param>
-        public void getValue(int[] data, SolidBrush brush)
+        public void getValue(float[] data, SolidBrush brush)
         {
-            radius = data[0];
+            radius = (int)data[0];
             b = brush;
         }
 
-/// <summary>
-/// This methhod takes the graphics int for initial position x and y bollen for drawing the shapes into shapes into the picture box.
-/// </summary>
-/// <param name="G"></param>
-/// <param name="x_axis"></param>
-/// <param name="y_axis"></param>
-/// <param name="turnfill"></param>
-    public void Draw (Graphics G , int x_axis , int y_axis, bool turnfill)
+        /// <summary>
+        /// This methhod takes the graphics int for initial position x and y bollen for drawing the shapes into shapes into the picture box.
+        /// </summary>
+        /// <param name="G"></param>
+        /// <param name="x_axis"></param>
+        /// <param name="y_axis"></param>
+        /// <param name="turnfill"></param>
+        public void Draw(Graphics G, int x_axis, int y_axis, bool turnfill)
         {
             if (turnfill)
             {
